OTWO-4172 Address Rubocop issue.           Move Amazon related methods to amazon.rb file           Fix accompanying tests           Increase bounce rate failure from 5% to 20% for initial testing
